requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiomysql==0.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
base58==2.1.1
bitarray==3.0.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.7
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
cryptography==44.0.0
cytoolz==1.0.1
et_xmlfile==2.0.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
fastapi==0.115.6
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hexbytes==1.3.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonalias==0.1.1
loguru==0.7.3
MarkupSafe==3.0.2
multidict==6.1.0
numpy>=1.24,<2.1
openpyxl==3.1.5
pandas==2.2.3
parsimonious==0.10.0
priority==2.0.0
propcache==0.2.1
psutil==7.0.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
Quart==0.19.9
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
solana==0.36.1
solders==0.23.0
SQLAlchemy==2.0.36
starlette==0.41.3
toolz==1.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3>=2.0.0,<3.0.0
utils==1.0.2
web3==7.8.0
websockets==12.0
Werkzeug==3.1.3
wheel==0.45.1
win32_setctime==1.2.0
wsproto==1.2.0
yarl==1.18.3